refactor(channel): extract header queries + state queries
Phase 2 continuation. Two pure-read extractions:

(1) channel_header.go (224 LOC, new):
    currentHeaderTitle, currentHeaderMeta, currentAppLabel,
    currentMainLines. Build the header strip strings + fetch
    cached main-viewport lines.

(2) channel_state_queries.go (86 LOC, new):
    visiblePendingRequest, composerTargetLabel, recommendedOptionIndex,
    interviewOptionCount, selectedInterviewOption, nextFocus. Tiny
    read-only helpers the Update method calls to make local decisions.

  - cmd/wuphf/channel.go: -278 LOC (2857 -> 2579)
  - cmd/wuphf/channel_header.go: +224 LOC (new)
  - cmd/wuphf/channel_state_queries.go: +86 LOC (new)

Cumulative channel.go reduction across Phase 2:
  4807 -> 2579
  Net: -2228 LOC; channel.go shrunk by 46%.

No behavior change. cmd/wuphf tests pass.
